minor: mostly style changes
the one non-style change is to pass an `ArrowIPCBuffer` to
`SchemaFromIPCBuffer` instead of a `uintptr_t`. Otherwise, I moved
`ArrowIPCTableFunction::GetFunction()` to the end (since it depends on
other functions and nothing depends on it) and I made some spacing
changes

NOTE: The commit removing the duckdb submodule was accidentally skipped
during a rebase, so I am adding it to this commit.